requirements.txt
streamlit
python-dotenv
PyPDF2
requests
beautifulsoup4
langchain
openai
